requirements.txt
aiofiles==0.8.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==8.1.0
colorama==0.4.4
cssselect==1.1.0
ffmpeg==1.4
ffmpeg-python==0.2.0
Flask==2.1.0
future==0.18.2
greenlet==1.1.2
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
imageio-ffmpeg==0.4.5
itsdangerous==2.1.2
Jinja2==3.1.1
lxml==4.8.0
MarkupSafe==2.1.1
msgpack==1.0.2
multidict==5.2.0
numpy==1.21.4
priority==2.0.0
psutil==5.8.0
py-cord==2.0.0b5
pyasn1==0.4.8
pycord==0.1.1
pycparser==2.21
PyNaCl==1.4.0
python-dotenv==0.19.2
pytube==11.0.1
quart==0.17.0
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
w3lib==1.22.0
Werkzeug==2.1.0
wsproto==1.1.0
yarl==1.7.0
youtube-dl==2021.6.6
youtube-search==2.1.0
